OCPBUGS-52471: fix OOM during channel filtering.
By not checking that the key exists in the `incoming` map, we were
inadvertently adding a channelEntry multiple times to the queue. For
some operator channels with long number of entries and long chains of
replace/skips, this could cause OOM.

I have also removed the `sorted` array which was not being used for
anything and just consuming memory.